song-vocabulary/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
ollama==0.1.5
instructor==0.4.5
duckduckgo-search==3.9.6
pydantic==2.4.2
python-dotenv==1.0.0
httpx>=0.25.2,<0.26.0
